refactor(#3479): extract terminal-success watcher helpers from recovery_engine.rs
Verbatim move of the 6-symbol terminal-watcher/start-offset cluster into recovery_engine/terminal_watcher.rs (137 LoC); recovery_engine.rs prod LoC 3718 -> 3607, giant baseline ratchet lowered.
