+++ ARCHITECTURE.md
@@ -449,7 +449,8 @@ src/
 │   │   │   ├── jsonl_extract.rs
 │   │   │   ├── output_path_detect.rs
 │   │   │   ├── phase_policy.rs
-│   │   │   └── status_panel.rs
+│   │   │   ├── status_panel.rs
+│   │   │   └── terminal_watcher.rs
 │   │   ├── recovery_paths/
 │   │   │   ├── controller_cutover.rs
 │   │   │   ├── mod.rs
