+++ docs/agent-maintenance/change-surfaces.md
@@ -767,7 +767,11 @@
     its G1/G2 snapshots from `external_input_relay_lease(...).map(|l| l.generation)`;
     +62 from #3304: slash-command canonical prompt keys for `<command-*>` XML vs
     `/command args` dedupe, plus focused loop skill-expansion regressions).
-  - `src/services/discord/recovery_engine.rs` (3718 lines; #3479 r8 extracted the
+  - `src/services/discord/recovery_engine.rs` (3607 lines; #3479 item-2 extracted
+    the terminal-success watcher / recovery start-offset helper cluster into the
+    sub-1000-prod-LoC leaf module `recovery_engine/terminal_watcher.rs` (137) —
+    behaviour-preserving move, externally-called helpers re-imported byte-identical;
+    #3479 r8 extracted the
     pure output-path-detect, phase-policy, and jsonl-extract clusters into the
     sub-1000-prod-LoC leaf modules `recovery_engine/output_path_detect.rs` (177),
     `recovery_engine/phase_policy.rs` (120), and `recovery_engine/jsonl_extract.rs`
